moved acceptance test to a separate file
